+++ src/components/card/link-card.stories.tsx
@@ -20,7 +20,8 @@ export const Default: Story = {
 			title: "Sample Title",
 			description: "This is a sample description for the news item.",
 			href: "https://example.com",
-			badgeText: "Tech",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: false,
 	},
@@ -34,7 +35,8 @@ export const WithDeleteButton: Story = {
 			title: "News with Delete Button",
 			description: "This news item has a delete button visible.",
 			href: "https://example.com",
-			badgeText: "News",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: true,
 		deleteAction: fn(),
@@ -61,7 +63,8 @@ export const WithoutDescription: Story = {
 			key: "4",
 			title: "News without Description",
 			href: "https://example.com",
-			badgeText: "General",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: false,
 	},
@@ -77,7 +80,8 @@ export const LongTitle: Story = {
 			description:
 				"This is also a longer description to test how the component handles longer description text that might be truncated.",
 			href: "https://example.com",
-			badgeText: "Long Content",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: false,
 	},
@@ -91,7 +95,8 @@ export const InternalLink: Story = {
 			title: "Internal Link Example",
 			description: "This links to an internal page using Next.js routing.",
 			href: "/viewer",
-			badgeText: "Internal",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: false,
 	},
@@ -105,7 +110,8 @@ export const ExternalLink: Story = {
 			title: "External Link Example",
 			description: "This opens in a new tab with target=_blank.",
 			href: "https://nextjs.org",
-			badgeText: "External",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: false,
 	},
@@ -119,7 +125,8 @@ export const InvalidUrl: Story = {
 			title: "Invalid URL Example",
 			description: "This has an invalid URL and should fallback to home.",
 			href: "javascript:alert('xss')",
-			badgeText: "Invalid",
+			primaryBadgeText: "Tech",
+			secondaryBadgeText: "example.com",
 		},
 		showDeleteButton: false,
 	},
