+++ src/features/contents/actions/get-contents.test.ts
@@ -1,13 +1,16 @@
 import { beforeEach, describe, expect, test, vi } from "vitest";
-import { contentsQueryRepository } from "@/features/contents/repositories/contents-query-repository";
+import { contentsQueryRepository } from "@/infrastructures/contents/repositories/contents-query-repository";
 import { getContentsCount, getExportedContents } from "./get-contents";
 
-vi.mock("@/features/contents/repositories/contents-query-repository", () => ({
-	contentsQueryRepository: {
-		findMany: vi.fn(),
-		count: vi.fn(),
-	},
-}));
+vi.mock(
+	"@/infrastructures/contents/repositories/contents-query-repository",
+	() => ({
+		contentsQueryRepository: {
+			findMany: vi.fn(),
+			count: vi.fn(),
+		},
+	}),
+);
 
 vi.mock("@/utils/auth/session", () => ({
 	getSelfId: vi.fn().mockResolvedValue("test-user-id"),
