+++ src/features/images/actions/get-images.ts
@@ -1,7 +1,7 @@
 import { cache } from "react";
 import { PAGE_SIZE } from "@/constants";
 import type { Status } from "@/domains/common/types";
-import { imageQueryRepository } from "@/features/images/repositories/image-query-repository";
+import { imageQueryRepository } from "@/infrastructures/images/repositories/image-query-repository";
 import { getSelfId } from "@/utils/auth/session";
 
 export const getExportedImages = cache(async (page: number) => {
