+++ src/features/news/actions/add-news.ts
@@ -3,8 +3,8 @@ import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
 import { validateNews } from "@/domains/news/services/news-domain-service";
-import { newsCommandRepository } from "@/infrastructure/news/repositories/news-command-repository";
-import { newsQueryRepository } from "@/infrastructure/news/repositories/news-query-repository";
+import { newsCommandRepository } from "@/infrastructures/news/repositories/news-command-repository";
+import { newsQueryRepository } from "@/infrastructures/news/repositories/news-query-repository";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
 import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 import type { ServerAction } from "@/utils/types";
