+++ src/features/news/actions/delete-news.ts
@@ -2,7 +2,7 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
-import { newsCommandRepository } from "@/infrastructure/news/repositories/news-command-repository";
+import { newsCommandRepository } from "@/infrastructures/news/repositories/news-command-repository";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
 import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 import type { ServerAction } from "@/utils/types";
