+++ src/infrastructures/images/repositories/image-command-repository.ts
