+++ src/infrastructures/images/repositories/image-query-repository.test.ts
