+++ src/infrastructures/images/repositories/image-query-repository.ts
