+++ src/infrastructures/news/repositories/news-query-repository.ts
